howtouse, 리뷰 삭제 완료

+++ Travely_User/app/src/main/java/com/hyeran/android/tooc/model/mypage/ReviewSaveResponseData.kt
@@ -3,5 +3,5 @@ package com.hyeran.android.tooc.model.mypage
 data class ReviewSaveResponseData (
         var storeIdx : Int,
         var content : String,
-        var liked : Int
+        var liked : Long
 )
